Expose more explicit tree structure in Ops.
This commit changes TreeOps (and FullBinaryTreeOps) to
expose the labels as a type parameter. While this is
sometimes unnecessary for Bonsai itself, it makes
working with Bonsai in situations where the labels
are important much easier.

It also changes the core abstraction to use .foldNode
instead of .isLeaf, .isNode, .children, and casting.
This ends up being a lot nicer to work with, especially
for binary trees where leaves and branches have different
labels.

(Since trees are never that deep, we don't have to
worry about the non-tail recursion that .foldNode
introduces.)

Review by @tixxit.